MadrigalWeb

Documentation for MadrigalWeb.

MadrigalWeb.MadrigalWeb

MadrigalWeb.Experiment

MadrigalWeb.download_file

MadrigalWeb.filter_by_kindat

MadrigalWeb.MadrigalWeb — Module

A Julia API to access the Madrigal database.

Reference:

Madrigal Database Documentation

madrigalWeb: a python API to access the Madrigal database

MadrigalWeb.Experiment — Type

A class that encapsulates information about a Madrigal Experiment.

Similar to the MadrigalExperiment class in the madrigalWeb python module.

MadrigalWeb.download_file — Function

throw controls whether to throw or silently return nothing on request error

MadrigalWeb.filter_by_kindat — Method

filter_by_kindat(expFileList, kindat)

Returns a subset of the experiment files in expFileList whose kindat is found in kindat argument.